requirements.txt
customtkinter
azure-devops
msrest
python-dateutil
requests
